pipline/requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.10.0
pandas==2.2.2
pyarrow==17.0.0
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
clickhouse-driver==0.2.10
